# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for a Online Webbased Auction System",
        "title": "Online Webbased Auction System API",
        "contact": {
            "name": "Puremike",
            "url": "http://github.com/puremike",
            "email": "digitalmarketfy@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Returns the status of the application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Get health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.healthResponse"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Authenticates a user using email and password.\nUpon successful authentication, a short-lived **JWT (access token)** is set as an `HttpOnly` cookie named `jwt`.\nA long-lived **refresh token** is also set as an `HttpOnly` cookie named `refresh_token`.\nBoth cookies are crucial for maintaining user session and subsequent authenticated requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_puremike_online_auction_api_internal_models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Two HttpOnly cookies are set: 'jwt' (access token) and 'refresh_token' (refresh token).",
                        "schema": {
                            "type": "header"
                        }
                    },
                    "400": {
                        "description": "Bad Request - invalid input",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "security": [
                    {
                        "jwtCookieAuth": []
                    }
                ],
                "description": "Clears the user's authentication cookies, effectively logging them out.",
                "tags": [
                    "Users"
                ],
                "summary": "Logout User",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "Refreshes the JWT access token using a valid refresh token.\nIf the refresh token is valid, a new JWT is generated and set as an `HttpOnly` cookie.\nA valid refresh token must be provided as an `HttpOnly` cookie named `refresh_token`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Refresh JWT Token",
                "responses": {
                    "200": {
                        "description": "Token refreshed successfully",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Refresh token not found or invalid",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Failed to generate new token",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Create a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_puremike_online_auction_api_internal_models.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_puremike_online_auction_api_internal_models.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/{username}": {
            "get": {
                "security": [
                    {
                        "jwtCookieAuth": []
                    }
                ],
                "description": "Retrieves the profile of the user associated with the access token.\nAccess token must be provided as an `HttpOnly` cookie named `jwt`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get User Profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username of the user to retrieve profile for",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_puremike_online_auction_api_internal_models.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or expired token",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "github_com_puremike_online_auction_api_internal_models.CreateUserRequest": {
            "type": "object",
            "required": [
                "confirm_password",
                "email",
                "full_name",
                "location",
                "password",
                "username"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6
                },
                "location": {
                    "type": "string",
                    "maxLength": 42,
                    "minLength": 6
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 6
                }
            }
        },
        "github_com_puremike_online_auction_api_internal_models.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_puremike_online_auction_api_internal_models.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.healthResponse": {
            "type": "object",
            "properties": {
                "api_version": {
                    "type": "string"
                },
                "environment": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Use a valid JWT token. Format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "jwtCookieAuth": {
            "description": "JWT (JSON Web Token) access token, sent as an HttpOnly cookie.",
            "type": "apiKey",
            "name": "jwt",
            "in": "cookie"
        },
        "refreshTokenCookie": {
            "description": "Refresh token, sent as an HttpOnly cookie.",
            "type": "apiKey",
            "name": "refresh_token",
            "in": "cookie"
        }
    }
}